#41 refactor: useCategory 폴더 변경 및 카테고리 모달 props 변경
